[TWINFACES-313] refactor: update imports
