+++ src/entities/factory/components/resource-link/index.ts
@@ -1,0 +1,2 @@
+export * from "./resource-link";
+export * from "./tooltip";
